+++ topcoffea/json/lumi.json
@@ -1,5 +1,6 @@
 {
- "2016": 36.33,
+ "2016APV": 19.52,
+ "2016": 16.81,
  "2017": 41.48,
  "2018": 59.83
 }
